dapp/requirements.txt
asgiref==3.5.2
Django==4.1.3
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-js-asset==2.0.0
django-mptt==0.14.0
django-parler==2.3
django-parler-rest==2.2
djangorestframework==3.14.0
easy-thumbnails==2.8.3
Pillow==9.3.0
psycopg2-binary==2.9.5
pytz==2022.6
sqlparse==0.4.3
